Add common teacher attributes and actions into teacher class.
