+++ student.rb
@@ -1,4 +1,7 @@
+require_relative 'high_five'
+
 class Student
+  include HighFive
   attr_reader :age, :phase
   attr_accessor :name
 
@@ -8,15 +11,9 @@ class Student
     @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
-
   def set_phase(num)
-    response = ""
     if num == @phase
-      response += "I'm doing phase #{@phase} again because "
-      response += "I put my learning first. I'm gonna rock it!"
+      response = "I'm doing phase #{@phase} again because I put my learning first. I'm gonna rock it!"
     else
       response = "Oooh, phase #{num}. I hope I'm ready!"
     end
@@ -25,9 +22,6 @@ class Student
   end
 
   def learn_stuff
-    response = ""
-    response += "WHOA! I've never thought of it quite like that before. "
-    response += "Now I feel like a genius!"
-    response
+    "WHOA! I've never thought of it quite like that before. Now I feel like a genius!"
   end
 end
